+++ packages/rich-text-editor/src/RichTextEditor/schema.ts
@@ -3,7 +3,8 @@ import {
   nodes as coreNodes,
   marks as coreMarks,
 } from "@cultureamp/rich-text-toolkit"
-import { ToolbarItems } from "./RichTextEditor"
+import { TOOLBAR_CONTROLS } from "../constants"
+import { ToolbarItems, ToolbarControlTypes } from "../types"
 
 export const defaultNodes: NodeSpec = {
   doc: coreNodes.doc,
@@ -12,45 +13,52 @@ export const defaultNodes: NodeSpec = {
   hardBreak: coreNodes.hardBreak,
 }
 
-export const createSchemaFromControls = (
-  controls: ToolbarItems[] | undefined
-) => {
+export const createSchemaFromControls = (controls?: ToolbarItems[]) => {
+  if (!controls) {
+    return createSchema()
+  }
+
+  const namesFromControls = controls.reduce(
+    (acc: string[], c: ToolbarItems) => [...acc, c.name],
+    []
+  ) as ToolbarControlTypes[]
+  return createSchema(namesFromControls)
+}
+export const createSchemaWithAll = () => createSchema(TOOLBAR_CONTROLS)
+
+function createSchema(controls?: ToolbarControlTypes[]) {
   if (!controls) {
     return new Schema({
       nodes: defaultNodes,
     })
   }
 
-  const allControlNames: string[] = controls.reduce(
-    (acc: string[], c: ToolbarItems) => [...acc, c.name],
-    []
-  )
   const newNodes = { ...defaultNodes }
   const newMarks = {}
 
-  if (allControlNames.includes("bold")) {
+  if (controls.includes("bold")) {
     newMarks["strong"] = coreMarks.strong
   }
 
-  if (allControlNames.includes("italic")) {
+  if (controls.includes("italic")) {
     newMarks["em"] = coreMarks.em
   }
 
-  if (allControlNames.includes("underline")) {
+  if (controls.includes("underline")) {
     newMarks["underline"] = coreMarks.underline
   }
 
-  if (allControlNames.includes("bulletList")) {
+  if (controls.includes("bulletList")) {
     newNodes["bulletList"] = coreNodes.bulletList
     newNodes["listItem"] = coreNodes.listItem
   }
 
-  if (allControlNames.includes("orderedList")) {
+  if (controls.includes("orderedList")) {
     newNodes["orderedList"] = coreNodes.orderedList
     newNodes["listItem"] = coreNodes.listItem
   }
 
-  if (allControlNames.includes("link")) {
+  if (controls.includes("link")) {
     newMarks["link"] = coreMarks.link
   }