+++ packages/rich-text-editor/src/types.ts
@@ -1,3 +1,16 @@
+export type ToolbarControlTypes =
+  | "bold"
+  | "italic"
+  | "underline"
+  | "orderedList"
+  | "bulletList"
+  | "link"
+
+export interface ToolbarItems {
+  name: ToolbarControlTypes
+  group?: string
+}
+
 export type EditorContentArray = Array<{ [key: string]: any }>
 
 export type EditorRows =